refactor: 유틸함수 정리 및 빌드 오류 제거
