+++ src/components/wishlist/roomList.tsx
@@ -6,12 +6,12 @@ import clsx from "clsx";
 import dayjs from "dayjs";
 import relativeTime from "dayjs/plugin/relativeTime";
 
-import { formatRelativeTime } from "@/utils/dateFormatter";
+import { formatRelativeTime } from "@/utils/formatter/dateFormatter";
 import {
   getDisplayStatus,
   getDisplayType,
   getDistanceInKm,
-} from "@/utils/propertyFormatter";
+} from "@/utils/formatter/propertyFormatter";
 
 import { ListingCardProps } from "@/types/listingCard";
