+++ src/utils/formatter/consentTypeFormatter.ts
