+++ src/utils/formatter/detailFormatter.tsx
